fix templates url and security

+++ app/controllers/templates_controller.rb
@@ -4,19 +4,18 @@ class TemplatesController < ApplicationController
 
   # Display all templates per company
 	def index
-		@user = User.find(current_user.id)
     # ScaleRep Admin has the ability to choose templates by company
-		if @user.role == "scalerep"
-        @client_companies = scalerep_director_client_company()
+		if is_scalerep_admin
+        @client_companies = client_companies()
+        
         if params.has_key?(:client_company)
-            @company = ClientCompany.find_by(name: params["client_company"])
+            @company = getCompanyByName(params["client_company"])
         else
-				  @company = ClientCompany.find(@user.client_company_id)
+				  @company = current_user.client_company
 			  end
 		else
-			  @company = ClientCompany.find(@user.client_company_id)
+			  @company = current_user.client_company
 		end
-
 		@templates = @company.templates.order('created_at DESC').paginate(:page => params[:page], :per_page => 20)
 	end
 
@@ -25,7 +24,7 @@ class TemplatesController < ApplicationController
   # Company Id parameter passed in from index page
 	def new
 		@user = User.find(current_user.id)
-		@client_company = ClientCompany.find_by(id: params[:company_id])
+		@client_company = ClientCompany.find_by(api_key: params[:company_api_key])
 		@template = Template.new
 	end
 
@@ -52,31 +51,45 @@ class TemplatesController < ApplicationController
 
 	def destroy
 		  @template = Template.find(params[:id])
-   		if @template.destroy
+   		
+      if !is_scalerep_admin && @template.client_company != current_user.client_company
+        redirect_to templates_path, notice: 'You cannot access this template!'
+      end
+
+      if @template.destroy
   				redirect_to templates_path, notice: 'Template was successfully created.'
 		  end
    end
 
    	
-   	def edit
-   		 @template = Template.find(params[:id])
-   	end
+  def edit
+   	  @template = Template.find(params[:id])
+
+      if !is_scalerep_admin && @template.client_company != current_user.client_company
+        redirect_to templates_path, notice: 'You cannot access this template!'
+      end
+  end
 
 
-   	def update
-   		  @user = User.find(current_user.id)
-		    @client_company = ClientCompany.find_by(id: @user.client_company.id)
-		    @template = Template.find_by(id: params[:id])
-        respond_to do |format|
-          if @template.update(template_params)
-            format.html { redirect_to templates_path(), notice: 'Template was successfully updated' }
-            format.json { render :index, status: :created}
-          else
-            format.html { render :update }
-            format.json { render json: @template.errors, status: :unprocessable_entity }
-          end
+  def update
+	    
+      @template = Template.find_by(id: params[:id])
+
+      if !is_scalerep_admin && @template.client_company != current_user.client_company
+        redirect_to templates_path, notice: 'You cannot access this template!'
+      end
+
+
+      respond_to do |format|
+        if @template.update(template_params)
+          format.html { redirect_to templates_path(), notice: 'Template was successfully updated' }
+          format.json { render :index, status: :created}
+        else
+          format.html { render :update }
+          format.json { render json: @template.errors, status: :unprocessable_entity }
         end
-   	end
+      end
+  end
 
 
   private
@@ -84,5 +97,15 @@ class TemplatesController < ApplicationController
     def template_params
       params.require(:template).permit(:title, :body, :subject)
     end
+
+    def client_companies()
+      client_companies = ClientCompany.where("account_live = ?", true).pluck(:name)
+      return client_companies
+    end
+
+    def getCompanyByName(company_name)
+      return ClientCompany.find_by(name: company_name)
+    end
+
 end
